Rearrange utils scripts
